requirements.txt
python-decouple^=3.3
setuptools~=49.2.1